Update code style.
