+++ sofa-boot-project/sofa-boot-core/tracer-sofa-boot/src/main/java/com/alipay/sofa/tracer/boot/kafka/processor/KafkaConsumerFactoryPostProcessor.java
@@ -25,6 +25,7 @@ import org.springframework.kafka.core.ConsumerFactory;
  * KafkaConsumerFactoryPostProcessor.
  *
  * @author chenchen6   2020/9/3 22:18
+ * @since 3.9.1
  */
 public class KafkaConsumerFactoryPostProcessor implements BeanPostProcessor {
     @Override
